requirements.txt
openai==1.95.0
pandas==2.2.2
schedule==1.2.2
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
gspread==6.1.2
oauth2client==4.1.3
flask==3.0.3 